+++ Bom.lua
@@ -78,6 +78,12 @@ function BOM.Print(t)
   DEFAULT_CHAT_FRAME:AddMessage(BOM.Color("808080", L.CHAT_MSG_PREFIX) .. t)
 end
 
+---Print a text with "BomDebug: " prefix in the game chat window
+---@param t string
+function BOM.Dbg(t)
+  DEFAULT_CHAT_FRAME:AddMessage(tostring(GetTime()) .. " " .. BOM.Color("883030", L.CHAT_MSG_PREFIX) .. t)
+end
+
 function BOM.Color(hex, text)
   return "|cff" .. hex .. text .. "|r"
 end
@@ -219,10 +225,17 @@ function BOM.ScrollMessage(self, delta)
   self:ResetAllFadeTimes()
 end
 
+function BOM.SetForceUpdate(reason)
+  --if reason ~= nil then
+  --  BOM.Dbg("Set force update: " .. reason)
+  --end
+  BOM.ForceUpdate = true
+end
+
 -- Something changed (buff gained possibly?) update all spells and spell tabs
 function BOM.OptionsUpdate()
-  BOM.ForceUpdate = true
-  BOM.UpdateScan()
+  BOM.SetForceUpdate("OptionsUpdate")
+  BOM.UpdateScan("OptionsUpdate")
   BOM.UpdateSpellsTab()
   BOM.MyButtonUpdateAll()
   BOM.MinimapButton.UpdatePosition()
@@ -419,8 +432,8 @@ function BOM.OptionsInsertSpells()
   end
 
   BOM.UpdateSpellsTab()
-  BOM.ForceUpdate = true
-  BOM.UpdateScan()
+  BOM.SetForceUpdate("OptionsInsertSpells")
+  BOM.UpdateScan("OptionsInsertSpells")
 end
 
 ---ChooseProfile
@@ -441,8 +454,8 @@ function BOM.ChooseProfile (profile)
 
   BOM.ClearSkip()
   BOM.PopupDynamic:Wipe()
-  BOM.ForceUpdate = true
-  BOM.UpdateScan()
+  BOM.SetForceUpdate("ChooseProfile")
+  BOM.UpdateScan("ChooseProfile")
 end
 
 ---When BomCharacterState.WatchGroup has changed, update the buff tab text to show what's
@@ -687,8 +700,8 @@ function BOM.Init()
     { "spellbook", L["SlashSpellBook"], BOM.GetSpells },
     { "update", L["SlashUpdate"],
       function()
-        BOM.ForceUpdate = true
-        BOM.UpdateScan()
+        BOM.SetForceUpdate()
+        BOM.UpdateScan("Macro /bom update")
       end },
     { "updatespellstab", "", BOM.UpdateSpellsTab },
     { "close", L["SlashClose"], BOM.HideWindow },
@@ -743,21 +756,24 @@ local function Event_ADDON_LOADED(arg1)
           end)
 end
 
-local function Event_UNIT_POWER_UPDATE(arg1, arg2)
+local function Event_UNIT_POWER_UPDATE(unitTarget, powerType)
   --UNIT_POWER_UPDATE: "unitTarget", "powerType"
-  if arg2 == "MANA" and UnitIsUnit(arg1, "player") then
-    if (BOM.PlayerManaMax or 0) <= (UnitPower("player", 0) or 0) then
-      BOM.ForceUpdate = true
+  if powerType == "MANA" and UnitIsUnit(unitTarget, "player") then
+    local max_mana = BOM.PlayerManaMax or 0
+    local actual_mana = UnitPower("player", 0) or 0
+
+    if max_mana <= actual_mana then
+      BOM.SetForceUpdate(nil)
     end
   end
 end
 
 local function Event_GenericUpdate()
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate()
 end
 
 local function Event_Bag()
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate()
   BOM.WipeCachedItems = true
 
   if BOM.CachedHasItems then
@@ -767,7 +783,7 @@ end
 
 local function Event_SpellsChanged()
   BOM.GetSpells()
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate("Evt Spells Changed")
   BOM.SpellTabsCreatedFlag = false
   BOM.OptionsInsertSpells()
 end
@@ -776,7 +792,7 @@ local bom_in_party = IsInRaid() or IsInGroup()
 
 local function Event_PartyChanged()
   BOM.PartyUpdateNeeded = true
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate("Evt Party Changed")
 
   -- if in_party changed from true to false, clear the watch groups
   local in_party = IsInRaid() or IsInGroup()
@@ -791,21 +807,21 @@ end
 local function Event_UNIT_SPELLCAST_errors(unit)
   if UnitIsUnit(unit, "player") then
     BOM.CheckForError = false
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate()
   end
 end
 
 local function Event_UNIT_SPELLCAST_START(unit)
   if UnitIsUnit(unit, "player") and not BOM.PlayerCasting then
     BOM.PlayerCasting = "cast"
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate()
   end
 end
 
 local function Event_UNIT_SPELLCAST_STOP(unit)
   if UnitIsUnit(unit, "player") and BOM.PlayerCasting then
     BOM.PlayerCasting = nil
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate()
     BOM.CheckForError = false
   end
 end
@@ -813,14 +829,14 @@ end
 local function Event_UNIT_SPELLCHANNEL_START(unit)
   if UnitIsUnit(unit, "player") and not BOM.PlayerCasting then
     BOM.PlayerCasting = "channel"
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate()
   end
 end
 
 local function Event_UNIT_SPELLCHANNEL_STOP(unit)
   if UnitIsUnit(unit, "player") and BOM.PlayerCasting then
     BOM.PlayerCasting = nil
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate()
     BOM.CheckForError = false
   end
 end
@@ -835,7 +851,7 @@ end
 
 ---On combat start will close the UI window and disable the UI. Will cancel the cancelable buffs.
 local function Event_CombatStart()
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate("Evt Combat Start")
   BOM.DeclineHasResurrection = true
   BOM.AutoClose()
   if not InCombatLockdown() then
@@ -847,7 +863,7 @@ end
 
 local function Event_CombatStop()
   BOM.ClearSkip()
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate("Evt Combat Stop")
   BOM.DeclineHasResurrection = true
   BOM.AllowAutOpen()
 end
@@ -861,7 +877,7 @@ end
 
 local function Event_LoadingStop()
   BOM.LoadingScreenTimeOut = GetTime() + BOM.LOADING_SCREEN_TIMEOUT
-  BOM.ForceUpdate = true
+  BOM.SetForceUpdate("Evt Loading Stop")
 end
 
 ---Event_PLAYER_TARGET_CHANGED
@@ -896,8 +912,8 @@ local function Event_PLAYER_TARGET_CHANGED()
 
   if newName ~= BOM.SaveTargetName then
     BOM.SaveTargetName = newName
-    BOM.ForceUpdate = true
-    BOM.UpdateScan()
+    BOM.SetForceUpdate("PlayerTargetChanged")
+    BOM.UpdateScan("PlayerTargetChanged")
   end
 
 end
@@ -985,20 +1001,20 @@ function BOM.UpdateTimer()
 
   if BOM.MinTimer and BOM.MinTimer < GetTime() then
     --print("MINTIMER!")
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate("MinTimer")
   end
 
   if BOM.CheckCoolDown then
     local cdtest = GetSpellCooldown(BOM.CheckCoolDown)
     if cdtest == 0 then
       BOM.CheckCoolDown = nil
-      BOM.ForceUpdate = true
+      BOM.SetForceUpdate("CheckCooldown")
     end
   end
 
   if BOM.ScanModifier and bom_last_modifier ~= IsModifierKeyDown() then
     bom_last_modifier = IsModifierKeyDown()
-    BOM.ForceUpdate = true
+    BOM.SetForceUpdate("ModifierKeyDown")
   end
 
   --
@@ -1018,8 +1034,10 @@ function BOM.UpdateTimer()
     bom_last_update_timestamp = GetTime()
     bom_fps_check = debugprofilestop()
 
-    BOM.UpdateScan()
+    BOM.UpdateScan("Timer")
 
+    -- If updatescan call above took longer than 6 ms, and repeated update, then
+    -- bump the slow alarm counter, once it reaches 6 we consider throttling
     if (debugprofilestop() - bom_fps_check) > 6 and BOM.RepeatUpdate then
       bom_slow_count = bom_slow_count + 1
 
@@ -1040,7 +1058,8 @@ end
 
 BOM.PlayerBuffs = {}
 
----UnitAura
+---Handles UnitAura WOW API call.
+---For spells that are tracked by Buffomat the data is also stored in BOM.PlayerBuffs
 ---@param unitId string
 ---@param buffIndex number Index of buff/debuff slot starts 1 max 40?
 ---@param filter string Filter string like "HELPFUL", "PLAYER", "RAID"... etc
@@ -1062,8 +1081,10 @@ function BOM.UnitAura(unitId, buffIndex, filter)
 
       if duration > 0 and (expirationTime == nil or expirationTime == 0) then
         local destName = (UnitFullName(unitId))
+
         if BOM.PlayerBuffs[destName] and BOM.PlayerBuffs[destName][name] then
           expirationTime = BOM.PlayerBuffs[destName][name] + duration
+
           if expirationTime <= GetTime() then
             BOM.PlayerBuffs[destName][name] = GetTime()
             expirationTime = GetTime() + duration
@@ -1096,7 +1117,7 @@ local function Event_COMBAT_LOG_EVENT_UNFILTERED()
       --BOM.PlayerBuffs[destName]=nil -- problem with hunters and fake-deaths!
       --additional check in bom_get_party_members
       --print("dead",destName)
-      BOM.ForceUpdate = true
+      BOM.SetForceUpdate("Evt UNIT_DIED")
 
     elseif BOM.SharedState.Duration[spellName] then
       if bit.band(sourceFlags, COMBATLOG_OBJECT_AFFILIATION_MINE) > 0 then
@@ -1186,8 +1207,8 @@ function BOM.HideWindow()
     if BOM.WindowVisible() then
       BomC_MainWindow:Hide()
       autoHelper = "KeepClose"
-      BOM.ForceUpdate = true
-      BOM.UpdateScan()
+      BOM.SetForceUpdate("HideWindow")
+      BOM.UpdateScan("HideWindow")
     end
   end
 end
@@ -1214,8 +1235,8 @@ function BOM.ToggleWindow()
   if BomC_MainWindow:IsVisible() then
     BOM.HideWindow()
   else
-    BOM.ForceUpdate = true
-    BOM.UpdateScan()
+    BOM.SetForceUpdate("ToggleWindow")
+    BOM.UpdateScan("ToggleWindow")
     BOM.ShowWindow()
   end
 end